feat: devops-scm-api语言改成kotlin #30

+++ devops-scm-provider/devops-scm-provider-git/devops-scm-provider-tgit/src/main/kotlin/com/tencent/devops/scm/provider/git/tgit/TGitPullRequestService.kt
@@ -93,7 +93,7 @@ class TGitPullRequestService(
                     .targetBranch(input.targetBranch)
 
             if (input.targetRepo is Long) {
-                builder.targetProjectId(input.targetRepo)
+                builder.targetProjectId(input.targetRepo as Long)
             }
 
             val mergeRequest = tGitApi.mergeRequestApi.createMergeRequest(repo.projectIdOrPath, builder.build())
